Moved Wrapper class out of __main__
Moved the Wrapper class out of __main__.py and into core.wrapper.
Implemented version.py to dynamically create and register our version
numbers.
