+++ wrapper/__init__.py
@@ -1,0 +1,5 @@
+from utils.version import get_version
+
+VERSION = (1, 0, 0, 'rc', 1)
+
+__version__ = get_version(VERSION)
